fix: added an alert while adding empty task
