+++ package.json
@@ -18,6 +18,7 @@
     "next": "14.2.5",
     "react": "^18",
     "react-dom": "^18",
+    "react-hot-toast": "^2.4.1",
     "simplebar-react": "^3.2.6",
     "styled-components": "^6.1.11"
   },
