+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
+import Todo from "@/components/todo";
 import { Box } from "@mui/material";
-import Todo from "./todo";
 
 export default function Home() {
   return (
